src/spacy/requirements.txt
cfenv
Faker
Flask
Flask-WTF
gunicorn
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-psycopg2
psycopg2-binary
scikit-learn
spacy
uvicorn
fastapi